rename some packages and classes
