+++ grakn-engine/src/main/java/ai/grakn/engine/attribute/deduplicator/AttributeDeduplicatorDaemon.java
@@ -16,14 +16,14 @@
  * along with this program.  If not, see <https://www.gnu.org/licenses/>.
  */
 
-package ai.grakn.engine.attribute.uniqueness;
+package ai.grakn.engine.attribute.deduplicator;
 
 import ai.grakn.GraknConfigKey;
 import ai.grakn.Keyspace;
 import ai.grakn.concept.ConceptId;
 import ai.grakn.engine.GraknConfig;
-import ai.grakn.engine.attribute.uniqueness.queue.Attribute;
-import ai.grakn.engine.attribute.uniqueness.queue.RocksDbQueue;
+import ai.grakn.engine.attribute.deduplicator.queue.Attribute;
+import ai.grakn.engine.attribute.deduplicator.queue.RocksDbQueue;
 import ai.grakn.engine.factory.EngineGraknTxFactory;
 import ai.grakn.kb.internal.EmbeddedGraknTx;
 import org.slf4j.Logger;
@@ -36,7 +36,7 @@ import java.util.Set;
 import java.util.concurrent.CompletableFuture;
 import java.util.stream.Collectors;
 
-import static ai.grakn.engine.attribute.uniqueness.AttributeDeduplicator.deduplicate;
+import static ai.grakn.engine.attribute.deduplicator.AttributeDeduplicator.deduplicate;
 
 /**
  * This class is responsible for de-duplicating attributes. It is done to ensure that every attribute in Grakn stays unique.
