+++ grakn-engine/src/main/java/ai/grakn/engine/attribute/deduplicator/queue/Attribute.java
@@ -16,7 +16,7 @@
  * along with this program.  If not, see <https://www.gnu.org/licenses/>.
  */
 
-package ai.grakn.engine.attribute.uniqueness.queue;
+package ai.grakn.engine.attribute.deduplicator.queue;
 
 import ai.grakn.Keyspace;
 import ai.grakn.concept.ConceptId;
