+++ grakn-engine/src/main/java/ai/grakn/engine/attribute/deduplicator/queue/QueueException.java
@@ -16,7 +16,7 @@
  * along with this program.  If not, see <https://www.gnu.org/licenses/>.
  */
 
-package ai.grakn.engine.attribute.uniqueness.queue;
+package ai.grakn.engine.attribute.deduplicator.queue;
 
 /**
  * TODO
